feat: files and responsibilities abstraction
