+++ src/config/database.js
@@ -1,0 +1,4 @@
+const monk = require('monk');
+const db = monk(process.env.MONGODB_URI);
+
+module.exports = db;
