+++ src/middleware/rateLimit.js
@@ -1,0 +1,6 @@
+const rateLimit = require('express-rate-limit');
+
+module.exports = rateLimit({
+  windowMs: 30 * 1000,
+  max: 1,
+})
