+++ src/middleware/requestValidation.js
@@ -1,0 +1,8 @@
+const yup = require('yup');
+
+const schema = yup.object().shape({
+  slug: yup.string().trim().matches(/^[\w\-]+$/i),
+  url: yup.string().trim().url().required(),
+});
+
+module.exports = schema;
